requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cleanup==9.0.0
django-filter==24.3
django-resized==1.0.3
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
uritemplate==4.1.1
